bug(fix-eslint-errors):  fix singleRecipe component
refactor component for eslint errors
connect component to store
[#154595824]
