+++ client/src/js/store/actions/actionConstants.js
@@ -1,0 +1,2 @@
+export const NEW_RECIPE_CREATED = 'NEW_RECIPE_CREATED';
+
